+++ R/get_match_lineups.R
@@ -80,8 +80,15 @@ get_match_lineups <- function(match_url) {
           .clean_match_advanced_stats_data() %>%
           dplyr::filter(!is.na(.data$Player_Num)) %>%
           dplyr::bind_cols(Team=team, Home_Away=home_or_away, .) %>%
-          dplyr::select(.data$Team, .data$Home_Away, .data$Player, .data$Player_Num, .data$Nation, .data$Pos, .data$Age, .data$Min, .data$Gls, .data$Ast, .data$CrdY, .data$CrdR) %>%
           dplyr::mutate(Player_Num = as.character(.data$Player_Num))
+
+        if(any(grepl("Nation", colnames(additional_info)))) {
+         additional_info <- additional_info %>%
+           dplyr::select(.data$Team, .data$Home_Away, .data$Player, .data$Player_Num, .data$Nation, .data$Pos, .data$Age, .data$Min, .data$Gls, .data$Ast, .data$CrdY, .data$CrdR)
+        } else {
+          additional_info <- additional_info %>%
+            dplyr::select(.data$Team, .data$Home_Away, .data$Player, .data$Player_Num, .data$Pos, .data$Age, .data$Min, .data$Gls, .data$Ast, .data$CrdY, .data$CrdR)
+        }
 
 
         lineup <- lineup %>%